Refactored the HTTP access logging a bit.
Moved the new stuff into the old (similar) filter, renamed that old
filter to be a better fit.

Tweaked a couple log entries.

BLUEBUTTON-1177
