+++ bluebutton-server-app/src/main/webapp/WEB-INF/web.xml
@@ -6,24 +6,14 @@
 		<realm-name>bluebutton-data-server</realm-name>
 	</login-config>
 
-	<!-- Adds some additional context variables that can be included in the 
-		app's logging. -->
+	<!-- Manage the MDC context and ensure that the app's NDJSON-formatted HTTP 
+		access log is populated. -->
 	<filter>
-		<filter-name>LoggingContextFilter</filter-name>
-		<filter-class>gov.hhs.cms.bluebutton.server.app.LoggingContextFilter</filter-class>
+		<filter-name>RequestResponseLoggingFilter</filter-name>
+		<filter-class>gov.hhs.cms.bluebutton.server.app.RequestResponseLoggingFilter</filter-class>
 	</filter>
 	<filter-mapping>
-		<filter-name>LoggingContextFilter</filter-name>
-		<url-pattern>/*</url-pattern>
-	</filter-mapping>
-
-	<!-- Produces a structured access log for the application. -->
-	<filter>
-		<filter-name>StructuredAccessLogFilter</filter-name>
-		<filter-class>gov.hhs.cms.bluebutton.server.app.StructuredAccessLogFilter</filter-class>
-	</filter>
-	<filter-mapping>
-		<filter-name>StructuredAccessLogFilter</filter-name>
+		<filter-name>RequestResponseLoggingFilter</filter-name>
 		<url-pattern>/*</url-pattern>
 	</filter-mapping>
